+++ src/Models/AuthenticationLogRecord.php
@@ -3,6 +3,7 @@
 namespace LangleyFoxall\LaravelAuthenticationLog\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use LangleyFoxall\LaravelAuthenticationLog\ConfigManagers\Omissions;
 
 class AuthenticationLogRecord extends Model
 {
@@ -17,5 +18,13 @@ class AuthenticationLogRecord extends Model
     public function authenticatable()
     {
         return $this->morphTo();
+    }
+
+    public static function createWithOmissions(array $data)
+    {
+        if(array_key_exists('credentials', $data)) {
+            $data['credentials'] = Omissions::omitCredentials($data['credentials']);
+        }
+        self::create(Omissions::omitFields($data));
     }
 }