+++ src/Subscribers/AuthenticationLogSubscriber.php
@@ -9,20 +9,18 @@ class AuthenticationLogSubscriber
 {
     public function handleAuthenticatableLogin($event) 
     {
-        $fields = [
+        AuthenticationLogRecord::createWithOmissions([
             'authenticatable_id' => $event->user->id,
             'authenticatable_type' => get_class($event->user),
             'eventType' => get_class($event),
             'user_ip' => request()->getClientIp(),
             'recorded_at' => now()
-        ];
-        
-        AuthenticationLogRecord::create(Omissions::omitFields($fields));
+        ]);
     }
 
     public function handleAuthenticatableFailed($event) 
     {
-        AuthenticationLogRecord::create([
+        AuthenticationLogRecord::createWithOmissions([
             'credentials' => Omissions::omitCredentials($event->credentials),
             'eventType' => get_class($event),
             'user_ip' => request()->getClientIp(),
@@ -32,7 +30,7 @@ class AuthenticationLogSubscriber
 
     public function handleAuthenticatableLogout($event) 
     {
-        AuthenticationLogRecord::create([
+        AuthenticationLogRecord::createWithOmissions([
             'authenticatable_id' => $event->user->id,
             'authenticatable_type' => get_class($event->user),
             'eventType' => get_class($event),
@@ -43,7 +41,7 @@ class AuthenticationLogSubscriber
 
     public function handleAuthenticatableRegistered($event) 
     {
-        AuthenticationLogRecord::create([
+        AuthenticationLogRecord::createWithOmissions([
             'authenticatable_id' => $event->user->id,
             'authenticatable_type' => get_class($event->user),
             'eventType' => get_class($event),
@@ -53,8 +51,8 @@ class AuthenticationLogSubscriber
     }
     public function handleAuthenticatableLockout($event)
     {
-        AuthenticationLogRecord::create([
-            'credentials' => Omissions::omitCredentials($event->request->query()),
+        AuthenticationLogRecord::createWithOmissions([
+            'credentials' => $event->request->query(),
             'eventType' => get_class($event),
             'user_ip' =>  $event->request->getClientIp,
             'recorded_at' => now(),
@@ -63,7 +61,7 @@ class AuthenticationLogSubscriber
     
     public function handleAuthenticatablePasswordReset($event) 
     {   
-        AuthenticationLogRecord::create([
+        AuthenticationLogRecord::createWithOmissions([
                 'authenticatable_id' => $event->user->id,
                 'authenticatable_type' => get_class($event->user),
                 'eventType' => get_class($event),